Move point and TeamFight themend entities to respectable packages
