+++ ippon/cup_fight_permissions_test.py
@@ -4,7 +4,7 @@ import unittest
 import django.test
 from django.contrib.auth.models import User
 
-from ippon.models import TeamFight
+import ippon.team_fight.models as tfm
 import ippon.team.models as tem
 import ippon.tournament.models as tm
 from ippon.permissions import IsCupFightOwnerOrReadOnly
@@ -23,7 +23,7 @@ class TestCupFightPermissions(django.test.TestCase):
         self.t1 = tem.Team.objects.create(tournament=self.to, name='t1')
         self.t2 = tem.Team.objects.create(tournament=self.to, name='t2')
 
-        self.tf = TeamFight.objects.create(aka_team=self.t1, shiro_team=self.t2, tournament=self.to)
+        self.tf = tfm.TeamFight.objects.create(aka_team=self.t1, shiro_team=self.t2, tournament=self.to)
         self.permission = IsCupFightOwnerOrReadOnly()
         self.request = unittest.mock.Mock()
         self.view = unittest.mock.Mock()
