+++ ippon/fight_permissions_test.py
@@ -4,7 +4,7 @@ import unittest
 import django.test
 from django.contrib.auth.models import User
 
-from ippon.models import TeamFight
+import ippon.team_fight.models as tfm
 import ippon.team.models as tem
 import ippon.tournament.models as tm
 import ippon.player.models as plm
@@ -36,7 +36,7 @@ class TestFightPermissions(django.test.TestCase):
                                             birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=c)
         self.t2.team_members.create(player=self.p2)
 
-        self.tf = TeamFight.objects.create(aka_team=self.t1, shiro_team=self.t2, tournament=self.to)
+        self.tf = tfm.TeamFight.objects.create(aka_team=self.t1, shiro_team=self.t2, tournament=self.to)
         self.fight = self.tf.fights.create(aka=self.p1, shiro=self.p2)
         self.permission = IsFightOwnerOrReadOnly()
         self.request = unittest.mock.Mock()
