+++ ippon/group_fight_views_test.py
@@ -6,7 +6,8 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APIClient, APITestCase
 
-from ippon.models import TeamFight, GroupFight, GroupPhase, Group
+from ippon.models import GroupFight, GroupPhase, Group
+import ippon.team_fight.models as tfm
 import ippon.team.models as tem
 import ippon.tournament.models as tm
 import ippon.player.models as plm
@@ -48,8 +49,8 @@ class GroupFightViewTest(APITestCase):
                                             birthday=datetime.date(year=2001, month=1, day=1), sex=1, club_id=c)
         self.t4.team_members.create(player=self.p4)
 
-        self.tf1 = TeamFight.objects.create(aka_team=self.t1, shiro_team=self.t2, tournament=self.to)
-        self.tf2 = TeamFight.objects.create(aka_team=self.t3, shiro_team=self.t4, tournament=self.to)
+        self.tf1 = tfm.TeamFight.objects.create(aka_team=self.t1, shiro_team=self.t2, tournament=self.to)
+        self.tf2 = tfm.TeamFight.objects.create(aka_team=self.t3, shiro_team=self.t4, tournament=self.to)
 
         self.phase = GroupPhase.objects.create(name='phase', tournament=self.to, fight_length=3)
 
@@ -110,8 +111,8 @@ class GroupFightViewSetAuthorizedTests(GroupFightViewTest):
         self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
         with self.assertRaises(GroupFight.DoesNotExist):
             GroupFight.objects.get(pk=self.gf1.id)
-        with self.assertRaises(TeamFight.DoesNotExist):
-            TeamFight.objects.get(pk=self.gf1.team_fight.id)
+        with self.assertRaises(tfm.TeamFight.DoesNotExist):
+            tfm.TeamFight.objects.get(pk=self.gf1.team_fight.id)
 
     def test_delete_not_existing_group_fight_returns_bad_request(self):
         response = self.client.delete(reverse('groupfight-detail', kwargs={'pk': -5}))
