+++ ippon/team/models.py
@@ -2,7 +2,6 @@ from django.db import models
 
 import ippon.player.models as plm
 
-
 class Team(models.Model):
     name = models.CharField(max_length=100, blank=False)
     tournament = models.ForeignKey('Tournament', related_name='teams', on_delete=models.CASCADE)
@@ -13,6 +12,7 @@ class Team(models.Model):
 
 
 class TeamMember(models.Model):
+    #from ippon.player.models import Player
     player = models.ForeignKey('Player', on_delete=models.CASCADE, related_name='team_member')
     team = models.ForeignKey('Team', related_name='team_members', on_delete=models.CASCADE)
