+++ ippon/urls.py
@@ -6,7 +6,9 @@ from rest_framework.schemas import get_schema_view
 import ippon.authorization_views
 import ippon.club.views
 import ippon.player.views
+import ippon.point.views
 import ippon.team.views
+import ippon.team_fight.views
 import ippon.tournament.views
 from ippon import views
 
@@ -20,9 +22,9 @@ router.register(r'participations', ippon.tournament.views.TournamentParticipatio
 router.register(r'tournament_admins', ippon.tournament.views.TournamentAdminViewSet)
 router.register(r'club_admins', ippon.club.views.ClubAdminViewSet)
 router.register(r'teams', ippon.team.views.TeamViewSet)
-router.register(r'points', views.PointViewSet)
+router.register(r'points', ippon.point.views.PointViewSet)
 router.register(r'fights', views.FightViewSet)
-router.register(r'team_fights', views.TeamFightViewSet)
+router.register(r'team_fights', ippon.team_fight.views.TeamFightViewSet)
 router.register(r'groups', views.GroupViewSet)
 router.register(r'group_phases', views.GroupPhaseViewSet)
 router.register(r'group_fights', views.GroupFightViewSet)
